fixes to auth system, fixes to major bugs in modules/quotes.sh

+++ modules/quotes.sh
@@ -2,14 +2,18 @@
 # quotes.sh -- quotes module for miyoko's shellbot
 
 addquote () {
-	qtext=$(echo "'$1'")
-	ts=$(echo "@@@ added by $send_nick")
-	ts=$(echo $ts | sed -e 's/|/%/g;s/ /|/g')
-	echo "$qtext|$ts" >> etc/mod_quotes
-	quotes=$(cat etc/mod_quotes)
-	rm etc/mod_quotes ; touch etc/mod_quotes
-	echo "$quotes" >>etc/mod_quotes
-	msg $dest $send_nick, your quote was added.
+	if [ $(echo -n $1 | wc -c) == "1" ] || [ $(echo -n $1 | wc -c) == "" ] ; then
+		msg $dest $send_nick, your quote was rejected.
+	else	
+		qtext=$(echo "'$1'")
+		ts=$(echo "@@@ added by $send_nick")
+		ts=$(echo $ts | sed -e 's/|/%/g;s/ /|/g')
+		echo "$qtext|$ts" >> etc/mod_quotes
+		quotes=$(cat etc/mod_quotes)
+		rm etc/mod_quotes ; touch etc/mod_quotes
+		echo "$quotes" >>etc/mod_quotes
+		msg $dest $send_nick, your quote was added.
+	fi
 }
 
 listquotes () {
